# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CodeMaster API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all demo projects
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: object
                required: [projects]
                properties:
                  projects:
                    type: array
                    items:
                      $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects, admin]
      summary: Create a new project (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a single project by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    put:
      operationId: updateProject
      tags: [projects, admin]
      summary: Update a project (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "200":
          description: Project updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    delete:
      operationId: deleteProject
      tags: [projects, admin]
      summary: Delete a project (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutSuccess'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /projects/{id}/favorite:
    post:
      operationId: toggleFavoriteProject
      tags: [projects, users]
      summary: Toggle favorite status for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        "200":
          description: Favorite toggled
          content:
            application/json:
              schema:
                type: object
                required: [favorited]
                properties:
                  favorited:
                    type: boolean
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /user/favorites:
    get:
      operationId: getUserFavorites
      tags: [users]
      summary: Get the current user's favorite projects
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        "200":
          description: List of favorited projects
          content:
            application/json:
              schema:
                type: object
                required: [projects]
                properties:
                  projects:
                    type: array
                    items:
                      $ref: "#/components/schemas/Project"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /user/profile:
    put:
      operationId: updateUserProfile
      tags: [users]
      summary: Update the current user's profile (first/last name)
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUser'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /user/account:
    delete:
      operationId: deleteUserAccount
      tags: [users]
      summary: Delete the current user's account permanently
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        "200":
          description: Account deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutSuccess'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /user/inquiries:
    get:
      operationId: getUserInquiries
      tags: [users]
      summary: Get the current user's submitted inquiries/leads
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        "200":
          description: List of inquiries
          content:
            application/json:
              schema:
                type: object
                required: [inquiries]
                properties:
                  inquiries:
                    type: array
                    items:
                      $ref: "#/components/schemas/Lead"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Opaque session token — `Bearer <sid>`.
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string
  schemas:
...
    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl]
      properties:
        id:
          type: string
        email:
          type: [string, 'null']
          format: email
        firstName:
          type: [string, 'null']
        lastName:
          type: [string, 'null']
        profileImageUrl:
          type: [string, 'null']
...
    LogoutSuccess:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          const: true
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
    Project:
      type: object
      required: [id, name, category, industry, description, features, status, featured]
      properties:
        id:
          type: string
        name:
          type: string
        category:
          type: string
        industry:
          type: string
        description:
          type: string
        features:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [demo, realization, concept]
        featured:
          type: boolean
        thumbnailUrl:
          type: [string, 'null']
        demoUrl:
          type: [string, 'null']
        technologies:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
    CreateProjectRequest:
      type: object
      required: [name, category, industry, description, features, status]
      properties:
        name:
          type: string
        category:
          type: string
        industry:
          type: string
        description:
          type: string
        features:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [demo, realization, concept]
        featured:
          type: boolean
        thumbnailUrl:
          type: [string, 'null']
        demoUrl:
          type: [string, 'null']
        technologies:
          type: array
          items:
            type: string
    Lead:
      type: object
      required: [id, name, email, projectType, description, status, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: [string, 'null']
        company:
          type: [string, 'null']
        projectType:
          type: string
        budget:
          type: [string, 'null']
        deadline:
          type: [string, 'null']
        description:
          type: string
        status:
          type: string
          enum: [new, in_progress, quoted, closed_won, closed_lost]
        notes:
          type: [string, 'null']
        priority:
          type: [string, 'null']
        userId:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
...
    UpdateProfileRequest:
      type: object
      properties:
        firstName:
          type: [string, 'null']
        lastName:
          type: [string, 'null']
...